+++ tests/functional_tests/test_app.py
@@ -1,6 +1,5 @@
 import pytest
 from harlequin.tui import Harlequin
-from harlequin.tui.components import CodeEditor
 
 
 @pytest.mark.asyncio
@@ -27,8 +26,55 @@ async def test_select_1(app: Harlequin) -> None:
 @pytest.mark.asyncio
 async def test_query_formatting(app: Harlequin) -> None:
     async with app.run_test() as pilot:
-        editor = app.query_one(CodeEditor)
-        editor.text = "select\n\n1 FROM\n\n foo"
+        app.editor.text = "select\n\n1 FROM\n\n foo"
 
         await pilot.press("ctrl+@")  # alias for ctrl+`
-        assert editor.text == "select 1 from foo\n"
+        assert app.editor.text == "select 1 from foo\n"
+
+
+@pytest.mark.asyncio
+async def test_run_query_bar(app_small_db: Harlequin) -> None:
+    app = app_small_db
+    async with app.run_test() as pilot:
+        # initialization
+        bar = app.run_query_bar
+        assert bar.checkbox.value is False
+        assert bar.input.value == "500"
+        assert app.limit == 500
+
+        # query without any limit by clicking the button;
+        # dataset has 857 records
+        app.editor.text = "select * from drivers"
+        await pilot.click(bar.button.__class__)
+        await app.workers.wait_for_complete()
+        assert len(app.data) > 500
+
+        # apply a limit by clicking the limit checkbox
+        await pilot.click(bar.checkbox.__class__)
+        assert bar.checkbox.value is True
+        await pilot.click(bar.button.__class__)
+        await app.workers.wait_for_complete()
+        assert len(app.data) == 500
+
+        # type an invalid limit, checkbox should be unchecked
+        # and a tooltip should appear on hover
+        await pilot.click(bar.input.__class__)
+        await pilot.press("a")
+        assert bar.input.value == "a500"
+        assert app.limit == 500
+        assert bar.checkbox.value is False
+        assert bar.input.tooltip is not None
+
+        # type a valid limit
+        await pilot.press("backspace")
+        await pilot.press("delete")
+        await pilot.press("1")
+        assert bar.input.value == "100"
+        assert app.limit == 100
+        assert bar.checkbox.value is True
+        assert bar.input.tooltip is None
+
+        # run the query with a smaller limit
+        await pilot.click(bar.button.__class__)
+        await app.workers.wait_for_complete()
+        assert len(app.data) == 100